+++ docs/project/changelog.md
@@ -65,7 +65,13 @@ myst:
 - Upgraded `zengl` to 2.5.0 {pr}`4894`
 - Upgraded `sourmash` to 4.8.11 {pr}`4980`
 - Upgraded `scipy` to 1.13.1 {pr}`4719`, {pr}`5011`
+- Upgraded `gensim` to 4.3.3 {pr}`4719`
+- Upgraded `numpy` to 2.1.0 {pr}`4925`
+- Upgraded `RobotRaconteur` to 1.2.2 {pr}`4925`
 - Upgraded `scikit-image` to 0.24.0 {pr}`5003`
+- Upgraded `yt` to 4.3.1 {pr}`4925`
+- Upgraded `opencv-python` to 4.3.1 {pr}`4925`
+- Upgraded `gsw` to 3.6.19 {pr}`4925`
 - Added `casadi` 3.6.5 {pr}`4936`
 - Added `pyarrow` 17.0.0 {pr}`4950`
 - Added `rasterio` 1.13.10, `affine` 2.4.0 {pr}`4983`